Move lineIntersectsCircle from Trig to MathUtils
This method doesn't make use of any trig functions.
