+++ NAS2D/MathUtils.cpp
@@ -8,10 +8,41 @@
 // = Acknowledgment of your use of NAS2D is appreciated but is not required.
 // ==================================================================================
 #include "MathUtils.h"
+#include <algorithm>
 #include <stdexcept>
 
 
 namespace NAS2D {
+
+/**
+ * Determines if a given line intersects a given circle.
+ *
+ * \param	p	First point of a line segment.
+ * \param	q	Second point of a line segment.
+ * \param	c	Center point of a circle.
+ * \param	r	Radius of a circle.
+ */
+bool lineIntersectsCircle(const Point<int>& p, const Point<int>& q, const Point<int>& c, float r)
+{
+	float dx = static_cast<float>(q.x() - p.x());
+	float dy = static_cast<float>(q.y() - p.y());
+
+	float t = -(((p.x() - c.x()) * dx) + ((p.y() - c.y()) * dy)) / ((dx * dx) + (dy * dy));
+
+	t = std::clamp(t, 0.0f, 1.0f);
+
+	dx = (p.x() + (t * (q.x() - p.x()))) - c.x();
+	dy = (p.y() + (t * (q.y() - p.y()))) - c.y();
+	float rt = (dx * dx) + (dy * dy);
+
+	if (rt < (r * r))
+	{
+		return true;
+	}
+
+	return false;
+}
+
 
 /**
  * Basic integer division that rounds up to the nearest whole number.
